+++ src/internet/model/tcp-veno.cc
@@ -25,6 +25,7 @@
  */
 
 #include "tcp-veno.h"
+#include "ns3/tcp-socket-base.h"
 #include "ns3/log.h"
 
 namespace ns3 {
@@ -161,7 +162,7 @@ TcpVeno::IncreaseWindow (Ptr<TcpSocketState> tcb, uint32_t segmentsAcked)
    * target cwnd is throughput / minRtt
    */
   double tmp = m_baseRtt.GetSeconds () / m_minRtt.GetSeconds ();
-  targetCwnd = static_cast<uint32_t> (segCwnd * tmp);
+  targetCwnd = segCwnd * tmp;
   NS_LOG_DEBUG ("Calculated targetCwnd = " << targetCwnd);
   NS_ASSERT (segCwnd >= targetCwnd); // implies baseRtt <= minRtt
